requirements-test.txt
pytest==7.4.4
pytest-cov==6.2.1
pytest-mock==3.12.0
pytest-xdist==3.8.0
coverage>=7.5
allure-pytest==2.13.2
playwright==1.45.0